docs: ODSC 2019 blog update.
Adding some imports in module initialization.
