+++ depiction/models/uri/__init__.py
@@ -1,0 +1,6 @@
+"""Initialize uri module."""
+from .cache import FileSystemModel  # noqa
+from .cache import HTTPModel  # noqa
+from .cache import COSModel  # noqa
+from .rest_api import RESTAPIModel  # noqa
+from .rest_api import MAXModel  # noqa
