+++ depiction/models/uri/cache/__init__.py
@@ -1,0 +1,4 @@
+"""Initialize cache module."""
+from .file_system_model import FileSystemModel  # noqa
+from .http_model import HTTPModel  # noqa
+from .cos_model import COSModel  # noqa
