+++ depiction/models/uri/rest_api/__init__.py
@@ -1,0 +1,3 @@
+"""Initialize rest_api module."""
+from .rest_api_model import RESTAPIModel  # noqa
+from .max_model import MAXModel  # noqa
