+++ depiction/models/uri/rest_api/max_model.py
@@ -1,6 +1,5 @@
 """Abstract interface for MAX models."""
 import os
-from abc import abstractmethod
 
 from .rest_api_model import RESTAPIModel
 
@@ -34,47 +33,3 @@ class MAXModel(RESTAPIModel):
         self.metadata = self._request(
             method='get', endpoint=self.metadata_endpoint
         )
-
-    @abstractmethod
-    def _process_prediction(self, prediction):
-        """
-        Process json prediction response.
-
-        Args:
-            prediction (dict): json prediction response.
-
-        Returns:
-            np.ndarray: numpy array representing the prediction.
-        """
-        raise NotImplementedError
-
-    @abstractmethod
-    def _predict(self, sample, *args, **kwargs):
-        """
-        Run the model for inference on a given sample and with the provided
-        parameters.
-
-        Args:
-            sample (object): an input sample for the model.
-            args (list): list of arguments.
-            kwargs (dict): list of key-value arguments.
-
-        Returns:
-            a prediction for the model on the given sample.
-        """
-        raise NotImplementedError
-
-    def predict(self, sample, *args, **kwargs):
-        """
-        Run the model for inference on a given sample and with the provided
-        parameters.
-
-        Args:
-            sample (object): an input sample for the model.
-            args (list): list of arguments.
-            kwargs (dict): list of key-value arguments.
-
-        Returns:
-            a prediction for the model on the given sample.
-        """
-        return self._process_prediction(self._predict(sample, *args, **kwargs))
